requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.0.0
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0
yfinance>=0.2.40
python-dotenv>=1.0.0
signalrcore>=0.9.5
websocket-client>=1.6.0
